+++ resources/views/layouts/app.blade.php
@@ -31,6 +31,8 @@
     </main>
 </div>
 <!-- Scripts -->
+<script src="{{ mix('js/manifest.js') }}"></script>
+<script src="{{ mix('js/vendor.js') }}"></script>
 <script src="{{ mix('js/app.js') }}"></script>
 </body>
 </html>